Create Parliament
An important resolution is being discussed in the Parliament of Chefland. There are 
𝑁
N members present in the Parliament out of which 
𝑋
X members voted in favour of the resolution and the remaining voted against it.

According to the constitution of Chefland, a resolution is passed if and only if half or more than half the members present in the Parliament vote in favour of the resolution.

Determine if the resolution is passed or not.